+ Added the ability to generate fractional random numbers. + Added "Scroll to the end of solution output after solving" option. * Set the accuracy for fractional numbers on output to 2 decimail places. * Updated translations to reflect changes.

+++ src/tspmodel.h
@@ -68,7 +68,7 @@ private:
 	bool loadError(QDataStream::Status);
 	bool loadTSPT(QDataStream *);
 	bool loadZKT(QDataStream *);
-	int rand(int, int) const;
+	double rand(int, int) const;
 };
 
 #endif // TSPMODEL_H
